+++ count.h
@@ -1,5 +1,5 @@
 /*
- * $Id: count.h,v 1.2 2005/06/22 21:52:53 jon Exp $
+ * $Id: count.h,v 1.3 2024/12/12 23:56:02 jon Exp $
  *
  * Function to count the non-zero elements in a matrix
  *
@@ -8,6 +8,6 @@
 #ifndef included__count
 #define included__count
 
-extern u32 count(const char *matrix, const char *name);
+extern u64 count(const char *matrix, const char *name);
 
 #endif
